backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
boto3==1.34.162
pydantic==2.8.2
twilio>=9.0.0
razorpay>=1.4.2